Publish curated HASP release v0.1.44
This release export was generated from the canonical source repo.

Source-Repo: private-monorepo
Source-Commit: 18d7cb0a94c1896f9262e5dd4f42986ac4b37801
Source-Tag: v0.1.44

+++ apps/server/internal/audit/file_state_linux.go
@@ -15,8 +15,8 @@ func auditFileStateFromInfo(info os.FileInfo) fileState {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
 		state.ctimeSec = stat.Ctim.Sec
 		state.ctimeNsec = stat.Ctim.Nsec
-		state.dev = uint64(stat.Dev)
-		state.ino = uint64(stat.Ino)
+		state.dev = stat.Dev
+		state.ino = stat.Ino
 		state.cacheable = true
 	}
 	return state
